paper_writing_system/requirements.txt
# Core Framework
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# LLM & Agent Framework
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# UI
streamlit>=1.28.0

# Data & Processing
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# Export Formats
python-docx>=0.8.11

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0
